win: fix winapi anonymous union syntax
Remove use of the `DUMMYUNIONNAME` macro from our definitions of
internal winapi anonymous unions.  This macro was added to `winternl.h`
in Windows SDK v7.0a in combination with a conditional definition of
`DUMMYUNIONNAME` to either empty or `u` depending on compiler support.
We support no compilers that do not have anonymous unions, but we do
support compilers such as VS 2008 that complain about the presence of
the `DUMMYUNIONNAME` identifier on an anonymous union because their
winapi headers do not define the macro publicly.  Simply remove it.