macOS: avoid including macOS CoreServices header globally
In `unix/internal.h` we include it only for `AvailabilityMacros.h`, so
include that directly instead.  We already include `CoreServices` where
needed in implementation files.

Apple documents a newer `Availability.h` for iOS and OS X >= 10.6, but
our reason for including the above header is to detect old OS X versions.