Allow custom read/write handlers in streams
When working with e.g. mbedtls, all ssl necessary
calls can be added in thenormal tcp workflow,
except reading and writing the tcp socket. To read
from mbedtls tcp sockets we need to call the
mbedtls_ssl_write and mbedtls_ssl_read functions,
instead of the regular read and write.

This commit allows a developer to override the
standard c read/write calls inside uv__read and
uv__write with custom implementations. For
writing a write_func was added to uv_write_t and
for reading a read_func was added to the
uv_stream_t.