v2: remove suggested_size from alloc_cb
This commit removes the  parameter in
the allocation callback. It has been relatively underused,
and users are able to allocate as much data as they want.

The  is now contained within the documentation
for users to have a good default to go to.

simplify tests by removing malloc/free calls

update the documentation to include UNIX FIONREAD ioctl infos