test: fix long output buffering
The tests runners sometimes forward the test output. But when the test
output is too long (over 1024 bytes), the runner output gets corrupted.

This is because unfinished lines are written without "%.*s" causing
bad behavior as the buffer is not always nul terminated (it's not the
case on windows).

The issue can be seen with the `platform_output` test on windows.

This commit changes the print_lines function so it keeps track of new
lines and never write more data than it should.