unix: fail with ENAMETOOLONG in `uv_pipe_*`
Fail with ENAMETOOLONG when the name of a Unix socket exceeds
`sizeof(saddr.sun_path)`. Previously the path was just truncated,
which could result in nasty bugs, and even though that behaviour
has been always been around, it’s hard to imagine a situation in
which ending up with an incorrect path is better than not creating
a socket at all.

Ref: https://github.com/nodejs/node/pull/12601
Ref: https://github.com/nodejs/node/issues/12708