unix,win: add uv_dlopen_with_flags()
Not surprisingly, this new function is used to pass non-default flags to dlopen.
To keep backwards compatibility, uv_dlopen() is maintained, which calls
dlopen with RTLD_LAZY.

For example, this function will be useful when adding support for dlopen
flags in Node.js.

Note that uv_dlopen_with_flags() is a no-op in windows, which just
calls uv_dlopen().

Signed-off-by: Ezequiel Garcia <ezequiel@vanguardiasur.com.ar>