fix cmpxchgi and cmpxchgl type error on Solaris
atomic_cas_ptr is for compare-and-swap pointer addresses. so when
building 64bit:
1. `ptr` is used casted to a pointer value (4-byte into 8-byte).
2. atomic_cas_ptr reads 8-byte at `ptr` and returns

in the case of `uv_async_send`, if handle->pending is 0, cmpxchgi
actually returns the value of the 4-bytes past handle->pending,
causing uv__async_send to never be called.

the modified test-async-null-cb.c hangs at uv_run in this case.

@see https://docs.oracle.com/cd/E86824_01/html/E54766/atomic-cas-ptr-3c.html