aix: fix un-initialized pointer field in fs handle
In AIX, fs watch close call was corrupting memory in the v8
compiler, when used in node. The handle->dir_filename field
can be un-initialized, if the watch is initiated but no event
got fired. The uv_fs_event_stop was freeing this pointer as if
it was malloc'ed, leading to the crash.

Properly initialize handle->dir_filename to avoid this being garbage.