unix: add if_indextoname
if_indextoname(...) is a function defined in <net/if.h>

It is used to convert an IPv6 scope_id into an interface
identifier string such as %eth0 or %lo

There is an identical function in Windows, but
that has not been tested.

Adds documentation and an ASSERT to the relevant test case.

PR-URL: https://github.com/libuv/libuv/pull/1443