android: fix blocking recvmsg due to netlink bug
Code was assuming that the nl_pid was always equal to the process id. As
the netlink documentation says, "nl_pid is usually the PID of the process
owning the destination socket. However, nl_pid identifies a netlink
socket, not a process".

There are two cases in which this assumption breaks down. The first is
that some other part of the application has already bound a netlink socket
with a nl_pid equal to the PID of the process. The second is Android
version 5.0 where it appears that nl_pid is not always defaulted to the
PID of the process.

The result of this bad assumption is that netlink message responses to our
request are ignored due to having a bad nl_pid. We eventually run out of
messages on the netlink socket and block forever awaiting our response.