netbsd: Correct uv_exepath() on NetBSD
Setup proper sysctl(2) arguments in order to retrieve the executable name.

Use an intermediate buffer beacause if it is too short, the sysctl(2) call
will return error. With this intermediate buffer populate the outer one.

Note issue with too long file names on NetBSD-8(beta), this is caused by the
current implementation of vnode to path translator in the kernel. It does not
cache longer file names than 31 character ones.