netbsd: Implement NetBSD-compatible uv__tty_is_slave()
NetBSD as an extension returns with ptsname(3) and ptsname_r(3) the slave
device name for both descriptors, the master one and slave one.

Workaround this problem and verify the device major and compare it with
the pts driver. Major numbers for the master and the slave TTY are
machine-dependent. On amd64 they are respectively 6 (ptc) and 5 (pts).