Let uv_timer_again() fail when called on a non-repeating timer.
The existing documentation and implementation contradict on what this
function is supposed to do when called on non-repeating timers. The
documentation says the timer should be stopped, whereas the
implementation leaves the timer alone. Disallowing it in the first place
makes more sense.

While there, remove the unnecessary call to uv_timer_stop(). The
uv_timer_start() function already stops the timer for us.

Fixes: #1591