pipe: fixed uv_pipe_chmod for OSX Darwin
According to the fstat man page there is a bug
in fstat on OS X that fstat on a pipe.
The fstat() call had to be changed to a stat() call
in order to get the correct permissions.

Test cases have been modified to check the result
of uv_pipe_chmod to confirm the change actually worked.