Add two process flags to allow proper job management
UV_PROCESS_START_SUSPENDED is necessary on windows to move
a process before it has a chance to create children. It could be
generally useful to a caller who wants to query or manipulate the
process after creation, though it is currently only implemented for
windows.

UV_PROCESS_NOMANAGE deals with the one-off already in place for the
windows process code which adds a process to a global job object
whenever the _DETACHED flag is not specified. This current behavior
is undesirable in pre-windows-8 versions because once a child goes
into this global job it cannot be assigned to another, but using
the detached flag may not be the desired effect either.