CMake: Install LICENSE once
The license file is installed twice. Once where it should be installed and the second time into the pkg-config directory.
Among other files from other projects it might be confusing why there is a license file among with other files.
Additionally, that is normally not the right location for license files.
Guess, that is only a copy/pasting mistake.