build: CMake: set OUTPUT_NAME for uv_a
It is not clear to me really why there cannot be just a single "uv"
library, which could be driven through `BUILD_SHARED_LIBS` to be shared
(by default), or static.

This change therefore provides to be a minimal fix, resulting in an
output of `libuv.a`, instead of `libuv_a.a`.