win: implement reflink on ReFS
Windows Server has a reflink feature built onto the dedup-capable ReFS.
This is completely irrelevant for client users, but it would still be a
good thing to add given how Microsoft loves Nodejs now.

The FICLONE_FORCE mode is likely to fail since you have to pay for the
server thing to get it to work. Remove the check if the CI is capable.
(We still really need someone to test it.)

This PR is based on the git-lfs implementation.

Ref: https://github.com/git-lfs/git-lfs/pull/3790